rag_server/requirements.txt
fastapi
uvicorn
moondream
pillow
python-multipart
python-dotenv
langchain
langchain-mistralai
pydantic
torch
sentence-transformers
rank-bm25
numpy
ast